Specific image uploader with previews
Based off the new file uploader component, this is more highly tuned
for image uploaders.

For full functionality it requires
`Intervention\Image\ImageManagerStatic` however works enough without it.

Database columns marked for image storage need to be of data type TEXT
or JSON as variants are stored

Usage is similar to that of `uploadFileToDisk` however this accepts a
3rd optional parameter for image variants.

Example usage would be within the Model

```php
protected $casts [‘image’ => ‘array’];

public function setImageAttribute($value)
    {
         $attribute_name = "image";
         $disk = "uploads";
         $destination_path = "folder_1/subfolder_1";

         $this->uploadImageToDisk($value, $attribute_name, $disk,
$destination_path);
    }
```

Then the Controller

$this->crud->setColumnDetails('image', ['type' => 'upload_image']);

$this->crud->addField([
                'name'  => 'image',
                'label' => 'hoto',
                'type'  => 'upload_image',
                'upload' => true
            ]);

//TODO
- Collaborate to make sure all methods of processing are agreed as the
best for the end user
- StyleCI Compatibility